+++ lib/factory_bot/with/proxy.rb
@@ -4,31 +4,28 @@ module FactoryBot
   class With
     # An intermediate object to provide some notation combined with <code>method_missing</code>.
     # @example
-    #   class Foo
-    #     def foo(name = nil)
+    #   class Example
+    #     def foo(name = nil, ...)
     #       return FactoryBot::With::Proxy.new(self, __method__) unless name
     #
     #       name
     #     end
     #   end
     #
-    #   Foo.new.foo.bar #=> :bar
+    #   ex = Example.new
+    #   ex.foo.bar #=> :bar
     class Proxy < BasicObject
       # @param receiver [Object]
       # @param method [Symbol]
-      # @param preargs [Array]
-      def initialize(receiver, method, *preargs)
+      def initialize(receiver, method)
         @receiver = receiver
         @method = method
-        @preargs = preargs
       end
 
       # @!visibility private
       def respond_to_missing?(_method_name, _) = true
 
-      def method_missing(method_name, ...)
-        @receiver.__send__(@method, *@preargs, method_name, ...)
-      end
+      def method_missing(method_name, ...) = @receiver.__send__(@method, method_name, ...)
     end
   end
 end