+++ spec/factory_bot/with/proxy_spec.rb
@@ -3,17 +3,16 @@
 RSpec.describe FactoryBot::With::Proxy do
   let(:test_class) do
     Class.new do
-      def foo(prefix, name = nil, *args, **kwargs, &block)
-        return FactoryBot::With::Proxy.new(self, __method__, prefix) unless name
+      def foo(name = nil, *args, **kwargs, &block)
+        return FactoryBot::With::Proxy.new(self, __method__, *args, **kwargs, &block) unless name
 
-        [prefix, name, args, kwargs, block]
+        [name, args, kwargs, block]
       end
     end
   end
 
   it "forwards method calls to the receiver" do
-    expect(test_class.new.foo(12).bar(34, abc: "def") { 345 }).to match [
-      12,
+    expect(test_class.new.foo.bar(34, abc: "def") { 345 }).to match [
       :bar,
       [34],
       { abc: "def" },